Use priority insertion into statusBar.
As indicated in the status-bar documentation here
https://github.com/atom/status-bar#api

Rather than using appendLeft(), addLeftTile() is used and a priority for
the wrap indicator is given. The example code for status-bar uses a
priority of 100 and many Atom packages dealing with the status bar have
simply used that priority in their code. As such, I have picked a
priority of 150 so that the wrap indicator will hopefully appear to the
farthest right side of the elements on the left hand side of the side bar.